Add FreeBSD support
FreeBSD does things its own way, sometimes.. The sudoers file is found
in /usr/local/etc instead of /etc.

Add this distinction in the manifests and also add a default sudoers
file for FreeBSD so that it sets up sudo access for the wheel group, as
is the default on this OS.

Finally, ensure that FreeBSD servers install the 'sudo' package.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>